pyproject.toml
[project]
name = "coral-interface-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiohttp>=3.12.9",
    "camel-ai[model-platforms]==0.2.70",
    "httpx-sse>=0.4.0",
    "mcp==1.10.1",
    "python-dotenv>=1.1.1",
    "starlette>=0.47.0",
    "yarl>=1.20.0",
]
